separate strem info
